requirements.txt
tensorflow>=2.15
tensorflow-datasets>=4.9
numpy>=1.24
matplotlib>=3.7
pillow>=10.0
scikit-learn>=1.3
